Use size_t instead of int for IdleConnList connection counts
Removes need for one type cast.

+++ src/adaptation/icap/ServiceRep.h
@@ -169,8 +169,8 @@ private:
 
     /// Set the maximum allowed connections for the service
     void setMaxConnections();
-    /// The number of connections which excess the Max-Connections limit
-    int excessConnections() const;
+    /// The number of connections which exceed the Max-Connections limit
+    size_t excessConnections() const;
     /**
      * The available connections slots to the ICAP server
      \return the available slots, or -1 if there is no limit on allowed connections
